// src/models/user.tsx
export interface UserModel {
    id: string;
    gender: string;
    name: string;
    username: string;
    email: string;
    address: string;
    city: string;
    state: string;
    country: string;
    postcode: string;
    phone: string;
    cell: string;
    nat: string;
    picture: string;
}